refactor: move chain config to lstar fork spec
The chain and consensus configuration constants lived under
node/chain/config.py, but they are protocol parameters that client
teams must match, not node implementation details.

Several spec/ modules already imported them from node/, an inverted
dependency where the protocol-spec layer reached down into a specific
node implementation. SSZ container bounds (HISTORICAL_ROOTS_LIMIT,
MAX_ATTESTATIONS_DATA) and forkchoice/timing parameters belong with
the fork they define.

Move the file to spec/forks/lstar/config.py, scoping the constants to
the lstar fork in line with the per-fork registry pattern, and update
all import sites across src, tests, and packages/testing.
