+++ packages/testing/src/consensus_testing/test_fixtures/slot_clock.py
@@ -8,12 +8,12 @@ to coordinate block proposals and attestations.
 from typing import Any, ClassVar
 
 from lean_spec.node.chain.clock import Interval, SlotClock
-from lean_spec.node.chain.config import (
+from lean_spec.spec.forks import Slot
+from lean_spec.spec.forks.lstar.config import (
     INTERVALS_PER_SLOT,
     MILLISECONDS_PER_INTERVAL,
     SECONDS_PER_SLOT,
 )
-from lean_spec.spec.forks import Slot
 from lean_spec.spec.ssz import Uint64
 
 from .base import BaseConsensusFixture
