+++ src/lean_spec/node/node.py
@@ -22,7 +22,6 @@ from typing import Final
 from lean_spec.node.api import AggregatorController, ApiServer, ApiServerConfig
 from lean_spec.node.chain import SlotClock
 from lean_spec.node.chain.clock import Interval
-from lean_spec.node.chain.config import ATTESTATION_COMMITTEE_COUNT
 from lean_spec.node.chain.service import ChainService
 from lean_spec.node.metrics import registry as metrics
 from lean_spec.node.networking import NetworkService
@@ -44,6 +43,7 @@ from lean_spec.spec.forks import (
     ValidatorIndex,
     Validators,
 )
+from lean_spec.spec.forks.lstar.config import ATTESTATION_COMMITTEE_COUNT
 from lean_spec.spec.ssz import Bytes32, Uint64
 
 logger = logging.getLogger(__name__)
