+++ src/lean_spec/spec/forks/lstar/containers.py
@@ -15,8 +15,8 @@ from pydantic import Field
 from lean_spec.base import StrictBaseModel
 from lean_spec.config import LEAN_ENV
 from lean_spec.node.chain.clock import Interval
-from lean_spec.node.chain.config import HISTORICAL_ROOTS_LIMIT
 from lean_spec.spec.crypto.xmss.containers import PublicKey, Signature
+from lean_spec.spec.forks.lstar.config import HISTORICAL_ROOTS_LIMIT
 from lean_spec.spec.ssz import Boolean, ByteList512KiB, Bytes32, Bytes52, Container, SSZList, Uint64
 from lean_spec.spec.ssz.bitfields import BaseBitlist
