+++ tests/consensus/lstar/fc/test_checkpoint_sync.py
@@ -12,9 +12,9 @@ from consensus_testing import (
 )
 
 from lean_spec.node.chain.clock import Interval
-from lean_spec.node.chain.config import INTERVALS_PER_SLOT, SECONDS_PER_SLOT
 from lean_spec.spec.crypto.merkleization import hash_tree_root
 from lean_spec.spec.forks import Slot, ValidatorIndex
+from lean_spec.spec.forks.lstar.config import INTERVALS_PER_SLOT, SECONDS_PER_SLOT
 from lean_spec.spec.ssz import Bytes32
 
 pytestmark = pytest.mark.valid_until("Lstar")
