+++ tests/consensus/lstar/fc/test_gossip_aggregated_attestation_validation.py
@@ -12,8 +12,8 @@ from consensus_testing import (
 )
 
 from lean_spec.node.chain.clock import Interval
-from lean_spec.node.chain.config import GOSSIP_DISPARITY_INTERVALS
 from lean_spec.spec.forks import Slot, ValidatorIndex
+from lean_spec.spec.forks.lstar.config import GOSSIP_DISPARITY_INTERVALS
 from lean_spec.spec.ssz import Bytes32
 
 pytestmark = pytest.mark.valid_until("Lstar")
