+++ tests/lean_spec/spec/forks/lstar/forkchoice/test_validate_attestation.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 import pytest
 
 from lean_spec.node.chain.clock import Interval
-from lean_spec.node.chain.config import GOSSIP_DISPARITY_INTERVALS, INTERVALS_PER_SLOT
 from lean_spec.spec.crypto.merkleization import hash_tree_root
 from lean_spec.spec.forks import Checkpoint, Slot, ValidatorIndex
 from lean_spec.spec.forks.lstar import Store
+from lean_spec.spec.forks.lstar.config import GOSSIP_DISPARITY_INTERVALS, INTERVALS_PER_SLOT
 from lean_spec.spec.forks.lstar.containers import (
     Attestation,
     AttestationData,
